+++ backend/src/routes/characters.ts
@@ -54,10 +54,12 @@ const router = express.Router();
 router.get("/", async (req: Request, res: Response) => {
   const characters = await CharacterRepo.find();
   for (const character of characters) {
-    character.image = process.env.URL + "/characters/uploads/" + character.id;
+    character.image =
+      process.env.URL + "/characters/uploads/" + character.id + ".jpg";
   }
   res.send(characters);
 });
+router.use("/uploads", express.static(charactersDir));
 
 router.post(
   "/",
@@ -76,6 +78,7 @@ router.post(
 
     if (file) {
       sharp(file.path)
+        .resize(500)
         .jpeg()
         .toFile(path.join(charactersDir, id + ".jpg"))
         .then(() => {
@@ -94,7 +97,7 @@ router.post(
   }
 );
 
-router.get("/uploads/:characterId", async (req: Request, res: Response) => {
+/* router.get("/uploads/:characterId", async (req: Request, res: Response) => {
   if (req.params.characterId == null) {
     res.status(400).send("Character ID not provided");
     return;
@@ -118,7 +121,7 @@ router.get("/uploads/:characterId", async (req: Request, res: Response) => {
   }
 
   return res.sendFile(filePath);
-});
+}); */
 
 router.put(
   "/:id",