+++ frontend/components/Tag/SelectableTagList.tsx
@@ -1,53 +1,60 @@
-import { useContext, Dispatch, SetStateAction } from "react";
-import { Typography, MenuItem, Button } from "@mui/material";
+import { useContext, Dispatch, SetStateAction, useState } from "react";
+import { TextField, Autocomplete, Button } from "@mui/material";
 import { DataContext } from "../../pages/_app";
 import Tag from "../../interfaces/Tag";
 import TagLabel from "./TagLabel";
+import TagChip from "./TagChip";
 interface TagListProps {
   selectedTags: Tag[];
   setSelectedTags: Dispatch<SetStateAction<Tag[]>>;
-  setOpen: Dispatch<SetStateAction<boolean>>;
-  filter: string;
 }
 
-export default function SelectableTagList(props: TagListProps) {
+export default function SelectableTagList({
+  setSelectedTags,
+  selectedTags,
+}: TagListProps) {
   const { data } = useContext(DataContext);
+  const [inputValue, setInputValue] = useState("");
 
-  function handleClick(tag: Tag) {
-    if (props.selectedTags.includes(tag)) {
-      props.setSelectedTags(props.selectedTags.filter((t) => t.id !== tag.id));
-      return;
-    }
-
-    props.setSelectedTags([...props.selectedTags, tag]);
-  }
-
-  function tagInSelectedTags(tag: Tag) {
-    // Check by tag id
-    return props.selectedTags.some((t) => t.id === tag.id);
+  function handleDeleteTag(tag: Tag) {
+    setSelectedTags(selectedTags.filter((t) => t.id !== tag.id));
   }
 
   return (
     <>
-      {data.tags.filter((artist) =>
-        artist.name.toLowerCase().includes(props.filter.toLowerCase())
-      ).length == 0 && <Typography>No tags</Typography>}
-      {data.tags
-
-        .filter((removeTag) =>
-          removeTag.name.toLowerCase().includes(props.filter.toLowerCase())
-        )
-
-        .map((removeTag) => (
-          <MenuItem
-            key={removeTag.id}
-            selected={tagInSelectedTags(removeTag)}
-            onClick={() => handleClick(removeTag)}
-          >
-            <TagLabel tag={removeTag} />
-          </MenuItem>
-        ))}
-      <Button onClick={() => props.setOpen(false)}>Close</Button>
+      <Autocomplete
+        multiple
+        options={data.tags}
+        getOptionLabel={(tag) => tag.name}
+        value={selectedTags}
+        inputValue={inputValue}
+        onInputChange={(event, newInputValue) => {
+          setInputValue(newInputValue);
+        }}
+        onChange={(event, newValue) => {
+          setSelectedTags(newValue);
+        }}
+        renderInput={(params) => (
+          <TextField {...params} variant="outlined" label="Tags" />
+        )}
+        renderOption={(props, tag) => {
+          return (
+            <li {...props}>
+              <TagLabel tag={tag} />
+            </li>
+          );
+        }}
+        renderTags={(value) =>
+          value.map((tag) => (
+            <TagChip
+              key={tag.id}
+              tag={tag}
+              onDelete={() => handleDeleteTag(tag)}
+            />
+          ))
+        }
+        noOptionsText={<Button>Create tag {inputValue}</Button>}
+      />
     </>
   );
 }