+++ Changelog.md
@@ -4,11 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.6.2] - 2023-09-18
+
+### Fixed
+
+- Fix problem with CORS policy
+
+### Changed
+
+- Tag Select uses Autocomplete from Material UI
+- Character select and artist select now don't use a popper
+
 ## [0.6.1] - 2023-09-15
 
 ### Fixed
+
 - Removed cors
-- 
+
 ## [0.6] - 2023-09-15
 
 ### Added